+++ src/posts/routes/handlers/delete-post.handler.ts
@@ -2,6 +2,8 @@ import { Request, Response } from 'express';
 import { HttpStatus } from '../../../core/types/http-statuses';
 import { errorsHandler } from '../../../core/errors/error.handler';
 import { postsService } from '../../application/posts.service';
+import { postRepository } from '../../repositories/postRepository';
+
 
 export async function deletePostHandler(
   req: Request<{ id: string }>,
@@ -9,6 +11,11 @@ export async function deletePostHandler(
 ) {
   try {
     const id = req.params.id;
+    const post = await postRepository.findById(id);
+
+    if (!post) {
+      throw new Error('Post not exist');
+    }
     await postsService.delete(id);
 
     res.sendStatus(HttpStatus.NoContent);
